booking-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
stripe==11.3.0
xero-python==4.2.0
PyJWT==2.10.1
structlog==24.4.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prometheus-fastapi-instrumentator==7.0.2
python-dotenv==1.0.1
httpx==0.28.1
